+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Revert AppleALC 1.7.6 update back to 1.6.3
   - Resolves audio issues on certain Intel HDEF devices
   - Regression currently being investigated within AppleALC
+- Remove `Force Web Drivers` option
+  - Avoids accidental use of non-Metal Web Drivers on Kepler GPUs
 
 ## 0.5.2
 - Ventura Specific Updates: